wmutils now compiles correctly
The problem I was having is that libxcb wasn't being found, but it did
exist in the system.  But since it wasn't being found I assumed it
wasn't installed or something.  Turns out it's just in a path where
make wasn't expecting it to, so I just had the update the config.mk file
and now it works.